Fix specs for Authlogic tokens and new cookies behaviour
Rails 5 tweaked the behaviour of cookies in controller specs in such a
way that cookie deletions don't show up. To work around this problem
set expectations on the response cookies hash to ensure that an empty
cookie is being sent to the client. Also adjust the length of the tokens
to the new length of 20 characters supplied by Authlogic.